TINKERPOP-1612 Factored out all the adapters between old/new plugin system in the console.
